refactor: create new package image. Move AlignedImage and ImagePoints to this package
