+++ src/main/java/com/ds4h/controller/importController/ImportController.java
@@ -1,6 +1,6 @@
 package com.ds4h.controller.importController;
 
-import com.ds4h.model.imagePoints.ImagePoints;
+import com.ds4h.model.image.imagePoints.ImagePoints;
 import com.ds4h.model.pointManager.PointManager;
 import com.ds4h.model.util.MemoryController;
 import com.ds4h.model.util.projectManager.importProject.ImportProject;
