+++ src/main/java/com/ds4h/model/alignment/automatic/pointDetector/kazeDetector/KAZEDetector.java
@@ -1,7 +1,7 @@
 package com.ds4h.model.alignment.automatic.pointDetector.kazeDetector;
 
 import com.ds4h.model.alignment.automatic.pointDetector.PointDetector;
-import com.ds4h.model.imagePoints.ImagePoints;
+import com.ds4h.model.image.imagePoints.ImagePoints;
 import org.opencv.core.*;
 import org.opencv.features2d.*;
 import java.util.List;
