+++ src/main/java/com/ds4h/model/alignment/automatic/pointDetector/orbDetector/ORBDetector.java
@@ -1,7 +1,7 @@
 package com.ds4h.model.alignment.automatic.pointDetector.orbDetector;
 
 import com.ds4h.model.alignment.automatic.pointDetector.PointDetector;
-import com.ds4h.model.imagePoints.ImagePoints;
+import com.ds4h.model.image.imagePoints.ImagePoints;
 import ij.IJ;
 import org.opencv.core.*;
 import org.opencv.features2d.BFMatcher;
