+++ src/main/java/com/ds4h/view/mainGUI/MainMenuGUI.java
@@ -11,11 +11,10 @@ import com.ds4h.controller.importController.ImportController;
 import com.ds4h.controller.opencvController.OpencvController;
 import com.ds4h.controller.pointController.ConvertLutImageEnum;
 import com.ds4h.controller.pointController.PointController;
-import com.ds4h.model.alignedImage.AlignedImage;
 import com.ds4h.model.alignment.alignmentAlgorithm.AffineAlignment;
 import com.ds4h.model.alignment.alignmentAlgorithm.ProjectiveAlignment;
 import com.ds4h.model.alignment.alignmentAlgorithm.TranslationalAlignment;
-import com.ds4h.model.imagePoints.ImagePoints;
+import com.ds4h.model.image.imagePoints.ImagePoints;
 import com.ds4h.view.aboutGUI.AboutGUI;
 import com.ds4h.view.automaticAlignmentConfigGUI.AutomaticAlignmentConfigGUI;
 import com.ds4h.view.bunwarpjGUI.BunwarpjGUI;
@@ -28,7 +27,6 @@ import com.ds4h.view.mosaicSettingsGUI.MosaicSettingsGUI;
 import com.ds4h.view.outputGUI.AlignmentOutputGUI;
 import com.ds4h.view.standardGUI.StandardGUI;
 import ij.IJ;
-import ij.ImagePlus;
 
 import javax.swing.*;
 import javax.swing.event.MouseInputListener;
