+++ src/main/java/com/ds4h/view/mainGUI/PreviewListItem.java
@@ -1,7 +1,7 @@
 package com.ds4h.view.mainGUI;
 
 import com.ds4h.controller.pointController.PointController;
-import com.ds4h.model.imagePoints.ImagePoints;
+import com.ds4h.model.image.imagePoints.ImagePoints;
 import com.ds4h.view.displayInfo.DisplayInfo;
 import com.ds4h.view.pointSelectorGUI.PointSelectorGUI;
 import com.ds4h.view.util.ViewBag;
