+++ src/main/java/com/ds4h/view/pointSelectorGUI/PointSelectorCanvas.java
@@ -1,6 +1,6 @@
 package com.ds4h.view.pointSelectorGUI;
 
-import com.ds4h.model.imagePoints.ImagePoints;
+import com.ds4h.model.image.imagePoints.ImagePoints;
 import com.ds4h.view.standardGUI.StandardCanvas;
 import ij.gui.*;
 import org.opencv.core.Point;
