+++ src/main/java/com/ds4h/view/pointSelectorGUI/PointSelectorGUI.java
@@ -1,7 +1,7 @@
 package com.ds4h.view.pointSelectorGUI;
 
 import com.ds4h.controller.pointController.PointController;
-import com.ds4h.model.imagePoints.ImagePoints;
+import com.ds4h.model.image.imagePoints.ImagePoints;
 import com.ds4h.view.mainGUI.MainMenuGUI;
 import com.ds4h.view.util.ViewBag;
 import ij.gui.ImageWindow;
