+++ src/main/java/com/ds4h/view/reuseGUI/ReuseListItem.java
@@ -1,6 +1,6 @@
 package com.ds4h.view.reuseGUI;
 
-import com.ds4h.model.alignedImage.AlignedImage;
+import com.ds4h.model.image.alignedImage.AlignedImage;
 import com.ds4h.view.displayInfo.DisplayInfo;
 import com.ds4h.view.standardGUI.StandardGUI;
 import javax.swing.*;
